refactor filter.go, move group functions to group.go
